Finished the login and create account
Finished the gui design and function of the login page and create account page.,
Create account inserts an account into parents table and loads the balance with a zero. All fields are require for a parent to create an account. Create student function hasn't been compeleted and neither is the table function on teh create account page. All of this is in the fx project. New Project name is LunchHour.